refactor(bluez): Expose all devices from Bluez
During the development of `bt status`, the initial version of `Bluez` only exposed the connected devices from Bluez's D-Bus.

It is noticed that the devices that are read from Bluez's D-Bus can be used in many different ways:

- For `status`, `bt` needs to get the connected devices from Bluez.
- For `list-devices`, `bt` needs to get all devices from Bluez, and filter them based on user inputs (if provided).

Therefore, the way Bluez exposes devices are changed.
Instead of implementing different queries and structs for each "usage" of a Bluez device (e.g BluezConnectedDev, BluezTrustedDev), a single query is written to get all devices and map them to a struct called `BluezDev`.

This way, when Bluez exposes the lowest common denominator - a device - this allows all the "clients" (`bt` in this case) to manipulate the devices however they want.

Signed-off-by: acikgozb <acikgozb@proton.me>