feat(ls): Implement `bt list-devices`
`bt list-devices` (or `bt ls` with alias) is implemented to show the known devices to the host.

It is designed to work in a similar way to `nmcli`. Like `nmcli`, it can provide the output in two different ways:

- A pretty output (table format),
- A terse output (for scripting).

Also, the output is designed to be filterable by specifying the field keys (e.g. "alias,connected), like `nmcli`.

Signed-off-by: acikgozb <acikgozb@proton.me>