feat(scan): Implement `bt scan`
The subcommand `bt scan` (or `bt sc` with alias) is implemented through
the custom Bluez DBus client.

The scanned devices are shown in different formats based on the usage:

- A pretty (table) format, for users,
- A terse format, for scripting purposes.

The implementation for different formats is similar to `bt ls`, however
they are kept separate for now since the whole project is in WIP state.

Signed-off-by: acikgozb <acikgozb@proton.me>