api(connect): Change `bt connect` API
During the development of `bt scan`, it is realized that `bt connect` may require a different sets of arguments than `bt scan`.

Therefore, the API design of `bt connect` is changed. The ability to change the interactive scan format is removed completely.

Instead, it is changed to allow a simple grep through scanned devices' ALIAS and a non-interactive mode by providing the full device ALIAS for convenience.

Signed-off-by: acikgozb <acikgozb@proton.me>