refactor(scan): Remove higher-level Bluez::scan()
While the initial scan functionality of Bluez client worked for
`bt scan`, a design issue was noticed during the `bt connect`
implementation:

- The created device object paths are flushed when
`Bluez::stop_discovery()` is called. However, it is not known exactly
when they are removed. The documentation says that it happens
within 3 minutes, which is not specific enough.

- This creates a non-deterministic behavior during `bt connect`
implementation because during the interactive scan, `bt connect` is
designed to wait for the user input to proceed. Since users can choose
whenever they want, the selected device should not be flushed before the
interactive scan.

In order to fix this behavior, the higher-level `Bluez::scan()` is
removed completely.

Instead, the caller now handles when to start and stop the device
discovery.

Signed-off-by: acikgozb <acikgozb@proton.me>