feat(connect): Implement `bt connect`
`bt connect` (or `bt c` with alias) is implemented to allow users
to connect to their Bluetooth devices in a more convenient way than
`bluetoothctl` offers.

For convenience, it has two modes:

- A non-interactive mode for cases where the device alias is known and
users just want to connect them directly.

- An interactive mode for cases where the device is not known or it's
alias is not convenient to type.
In this mode, a device scan is initiated and the results are shown to
the user via a pretty format, similar to other `bt` commands.

Once the alias is known via one of the methods above, `bt connect` tries
to establish a connection through `org.bluez.Device1` interface of the
device.

Signed-off-by: acikgozb <acikgozb@proton.me>