feat(bluez): Move Bluez init to bin root
Instead of initializing Bluez inside the subcommand modules, it is initialized inside the bin root.

This is done to make the lib crate more testable and extensible.

For now, the `BluezClient` type is left as a concrete type, but it will be converted into a trait later on to satisfy the goal above.

Signed-off-by: acikgozb <acikgozb@proton.me>