feat(format): Add `format` module
A new module called `format` is added to centralize the formatting of each `bt` subcommand.

This module provides 3 different traits, which should be implemented by callers that wish to provide pretty/terse output.

TableFormattable - The main trait that converts a pair of types into columns and cells.

PrettyFormatter - The trait that provides pretty (table) output for every type that is TableFormattable.

TerseFormatter - The trait that provides script (terse) output for every type that is TableFormattable.

The module also provides blanket implementations for `PrettyFormatter` and `TerseFormatter` for iterators of `TableFormattable` types.

Blanket implementations allow callers to automatically provide pretty/terse output when they have an Iterator of TableFormattable types.

Signed-off-by: acikgozb <acikgozb@proton.me>