feat(disconn): Exit early if there are no conn devs
During the interactive mode of `disconnect`, it is noticed that if there
are no connected devices, the list is shown as empty and users are still
asked to choose a device.

This does not happen in non-interactive mode, in that case the Bluez
client returns the appropriate error `bluez::Error::InterfaceNotFound`.

Therefore, in order to prevent this incorrect state during the
interactive mode, a new error `disconnect::Error::NoConnectedDevices`
is added.

Signed-off-by: acikgozb <acikgozb@proton.me>