feat(bluez): Introduce BluezTestClient for testing
Up until this point, all of the subcommands were using the "real" Bluez
DBus client.

Since the functionalities need to be tested, the "real" Bluez DBus
client needs to be swapped with a different implementation for the
tests.

The usual way of doing this in other languages is to design the
subcommands around a single Bluez client interface, and mock the client
in the tests.

Whilst this is applicable in some cases, Rust's type system is so
strict, therefore this would require a change in the actual code to
use dynamic dispatch (e.g. `Box<impl Client>`), just to make the code
testable.

Instead of introducing abstractions just for testing purposes, a new
Bluez test client struct `BluezTestClient` is added in `bluez` module.

To use the struct in tests, the client export is updated to export
`BluezTestClient` for tests and `BluezDBusClient` for the actual flow.

Signed-off-by: acikgozb <acikgozb@proton.me>